+++ SampleSite/Program.cs
@@ -7,6 +7,19 @@ namespace SampleSite
         public static void Main(string[] args)
         {
             var builder = WebApplication.CreateBuilder(args);
+
+            // Configure cookie-based authentication.
+            builder.Services
+                .AddAuthentication(CookieAuthenticationDefaults.AuthenticationScheme)
+                .AddCookie(o =>
+                {
+                    o.Cookie.HttpOnly = true;
+                    o.Cookie.SameSite = SameSiteMode.Strict;
+                    o.Cookie.SecurePolicy = CookieSecurePolicy.Always;
+                    o.ExpireTimeSpan = TimeSpan.FromMinutes(120);
+                    o.SlidingExpiration = true;
+                });
+
             builder.Services.AddControllersWithViews();
 
             var app = builder.Build();
@@ -16,6 +29,10 @@ namespace SampleSite
             }
             app.UseStaticFiles();
             app.UseRouting();
+
+            // Activate the cookie-based authentication.
+            app.UseAuthentication();
+
             app.UseAuthorization();
             app.MapControllerRoute(
                 name: "default",